Encrypt DNS requests of devices not using the VPN
I had decided to separate the DNS requests from the devices using the VPN gateway and the devices connecting directly to the internet because the with unbound DNS server is running locally and it's easily possible to figure out the IP the DNS server uses to contact the root nameservers. This would mean that the IP of the VPN connection and your own can be associated. However, this also means that for the device connected directly DNS requests were made directly to the root DNS servers (which don't support encryption). Thus, they could potentially be read and associated with you. With the extra layer the VPN provides they are now encrypted until they leave the VPN exit node and can't be associated with you anymore.